PROJECT FINISHED

+++ libft/ft_uitoa_base.c
@@ -6,23 +6,23 @@
 /*   By: ablaamim <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/05 17:48:17 by ablaamim          #+#    #+#             */
-/*   Updated: 2022/01/05 17:52:18 by ablaamim         ###   ########.fr       */
+/*   Updated: 2022/01/06 17:10:15 by ablaamim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	ft_convbase(unsigned long nbr, char *num, char *base, size_t i)
+static void	ft_convbase(unsigned long long nbr, char *num, char *base, size_t i)
 {
 	size_t	len_base;
 
 	len_base = ft_strlen(base);
-	if (nbr >= (unsigned long int)len_base)
+	if (nbr >= (unsigned long long)len_base)
 		ft_convbase((nbr / len_base), num, base, (i - 1));
 	num[i] = base[nbr % len_base];
 }
 
-static size_t	ft_countsize(unsigned long n, size_t len_base)
+static size_t	ft_countsize(unsigned long long n, size_t len_base)
 {
 	if ((n / len_base) == 0)
 		return (1);
@@ -30,7 +30,7 @@ static size_t	ft_countsize(unsigned long n, size_t len_base)
 		return (1 + ft_countsize(n / len_base, len_base));
 }
 
-char	*ft_uitoa_base(unsigned long nbr, char *base)
+char	*ft_uitoa_base(unsigned long long nbr, char *base)
 {
 	size_t		len_nbr;
 	size_t		len_base;
